remove util function, combine with API
